push for pull request, admin management done

+++ app/Http/Controllers/BooksController.php
@@ -43,21 +43,25 @@ class BooksController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'author' => 'required',
-            'title' => 'required',
-            'genre' => 'required',
-            'quantity' => 'required'
-        ]);
+        if(Auth::user()->isAdministrator()){
+            $request->validate([
+                'author' => 'required',
+                'title' => 'required',
+                'genre' => 'required',
+                'quantity' => 'required'
+            ]);
 
-        Books::create([
-            'author' => $request->author,
-            'title' => $request->title,
-            'genre' => $request->genre,
-            'quantity' => $request->quantity
-        ]);
-
-        return redirect()->route('books.index');
+            Books::create([
+                'author' => $request->author,
+                'title' => $request->title,
+                'genre' => $request->genre,
+                'quantity' => $request->quantity
+            ]);
+            return redirect()->route('books.index');
+        }
+        else{
+            abort(403);
+        }
     }
 
     /**
@@ -79,7 +83,12 @@ class BooksController extends Controller
      */
     public function edit(Books $book)
     {
-        return view('library.edit', compact('book'));
+        if(Auth::user()->isAdministrator()){
+            return view('library.edit', compact('book'));
+        }
+        else{
+            abort(403);
+        }
     }
 
     /**
@@ -91,21 +100,26 @@ class BooksController extends Controller
      */
     public function update(Request $request, Books $book)
     {
-        $request->validate([
-            'author' => 'required',
-            'title' => 'required',
-            'genre' => 'required',
-            'quantity' => 'required'
-        ]);
+        if(Auth::user()->isAdministrator()){
+            $request->validate([
+                'author' => 'required',
+                'title' => 'required',
+                'genre' => 'required',
+                'quantity' => 'required'
+            ]);
 
-        $book->update([
-            'author' => $request->author,
-            'title' => $request->title,
-            'genre' => $request->genre,
-            'quantity' => $request->quantity
-        ]);
+            $book->update([
+                'author' => $request->author,
+                'title' => $request->title,
+                'genre' => $request->genre,
+                'quantity' => $request->quantity
+            ]);
 
-        return redirect()->route('books.index');
+            return redirect()->route('books.index');
+        }
+        else{
+            abort(403);
+        }
     }
 
     /**
@@ -116,8 +130,13 @@ class BooksController extends Controller
      */
     public function destroy(Books $book)
     {
-        $book->delete();
+        if(Auth::user()->isAdministrator()){
+            $book->delete();
 
-        return redirect()->route('books.index');
+            return redirect()->route('books.index');
+        }
+        else{
+            abort(403);
+        }
     }
 }
